pyproject.toml
[project]
name = "shopping-list"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.129.2",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "ruff>=0.15.2",
    "sqlmodel>=0.0.37",
    "ty>=0.0.18",
    "uvicorn>=0.41.0",
]
